requirements.txt
RPi.GPIO
opencv-python
numpy
scipy
pylidar